# 73x times faster metrics correlations at the agent