… health/health.c
… libnetdata/storage_number/storage_number.h
… web/api/badges/web_buffer_svg.c
… web/api/formatters/json_wrapper.c
… web/api/formatters/json_wrapper.h
… web/api/formatters/rrd2json.c
… web/api/formatters/rrd2json.h
… web/api/netdata-swagger.json
… web/api/netdata-swagger.yaml
… web/api/queries/query.c
… web/api/queries/query.h
… web/api/queries/rrdr.c
… web/api/queries/rrdr.h
… web/api/web_api_v1.c
… web/api/web_api_v1.h
… web/server/web_client.h