This PR improves the performance of metrics correlations by a factor of 73.

The biggest problem of it was this function:

```c
int find_index(double arr[], long int n, double K, long int start)
{
    for (long int i = start; i < n; i++) {
        if (K<arr[i]){
            return i;
        }
    }
    return n;
}
```

It has been replaced with a binary search function, which alone saved 75% of CPU resources required.

Then the next big problem was comparisons of `double` numbers. So, I multiplied the `double` pair diffs by 1 million (to have 6 digits precision) and divided the final `min` and `max` by 1 million, so that we only compare `long int` numbers. This saved 15% more (compared to the original).

Then a number of additional optimizations have been done to optimize the whole logic further:

1. The need to replace NANs with zeros has been replaced with the option `RRDR_OPTION_NULL2ZERO` so that the query engine will do it.
2. Loops to copy data around have been eliminated
3. Loops to calculate min and max have been eliminated

The current code in the master, has another important issue. The baseline is 4x bigger than the highlight, but the code requests the same number of points for both the highlight and the baseline. This is now fixed. The new code tries to have the same duration for each point of the result. So, it first does the highlight query to find the number of points returned and then it does the baseline query requesting 4 times the number of points of the highlight. The code is smart enough to adapt the baseline window to be power of 2 compared to the highlight and respecting the maximum buffers the code has. So, the old code, also gives wrong results.

Overall, for a given test I made for 30 minutes of metrics correlations, by observations are like follows:

1. The original time to run metric correlations was about 8 seconds.
2. The time to do the queries to collect the data was about 700 ms (just 9% of the total time).
3. The new time to run metric correlations is about 800 ms (for the same timeframe, the same dataset).
4. So, 7300ms before vs 100ms now. 73x times faster.

Now metrics correlations is just about 12% more resource hungry than running metrics correlations on the cloud. It is really nothing. If we accumulated into this the CPU time the agent needs to prepare JSON responses for all data queries and send them over to the cloud, the agent version is now really a lot more efficient for the agent itself.

I also enabled metric correlations at the agent, by default.