@ktsaou 2022-06-09
@andrewm4894 you have to review this.

I found some discrepancies with the new and the old logic.
For example, the previous logic was giving about 2500 dimensions with a value of 1. Now the new logic gives only 250. All the others are the same. Just a lot of 1s became 0s. I really can't tell why.

Still this logic needs to be validated.