@andrewm4894 2022-06-10
Good stuff! I'll spin up some agents on this branch and do same comparison to the cloud results we did originally.