@ilyam8 2022-06-10
@andrewm4894 @MrZammler If that is technically possible I'd suggest adding some unit tests with input and **expected** correlations output. So we can improve our implementation and not be afraid that something is broken.