@ktsaou 2022-06-10
Now its 10x times faster!