@ktsaou 2022-06-10
So, the actual comparison is like this:

1. The queries need 700ms
2. The original time was 8 seconds. So 7300ms was metric correlations
3. Now total time is 800ms. Minus the queries (700ms), metric correlations is now 100ms.
4. So, the improvement is like 73 times faster...