@andrewm4894 2022-06-10
> @andrewm4894 @MrZammler If that is technically possible I'd suggest adding some unit tests with input and **expected** correlations output. So we can improve our implementation and not be afraid that something is broken.

Yeah it's a bit manual at the moment since have to test against cloud but going forward for sure we could have test data and expected outputs just for the endpoint which would be easy when testing iterations of the api and not needing to validate against cloud since we really only needed to do that for the original agent implementation. Will follow up with @MrZammler to make separate issues for some unit tests like you mention.