@ktsaou 2022-06-10
Difference on the agent compared to the cloud version, using a 48 minute highlighted area, running on RPI4b.

## Metric Correlations of this PR:

![image](https://user-images.githubusercontent.com/2662304/173075253-ea9d614a-d48d-4856-9054-b49768d87929.png)

## Metric Correlations of cloud version (impact on agent):

![image](https://user-images.githubusercontent.com/2662304/173075503-4e4dcef3-95e6-4f6e-979f-41ff7f680c3d.png)

As you can see, the CPU load is twice as high (2 cores used out of 4) and 13 seconds instead of 8 of the version above.
And the cloud does NOT work - does not bring results (I think it times out) !