@andrewm4894 2022-06-10
some initial testing of resources on identical machines, one on master one running this branch. 

I highlighted a 10 min window and ran MC on both agents. 

yellow is what's on master, blue is this branch.

![image](https://user-images.githubusercontent.com/2178292/173093934-b5ff1213-5366-41c5-a0e8-07635cbab84e.png)

i will test that the actual results returned are the same between both branches and vs cloud