@andrewm4894 2022-06-10
hmm - so if i use [this notebook](https://colab.research.google.com/drive/1DHx7_23JPQyZwLc-VtSmeKFirvJWjHD_?usp=sharing) to compare agent to cloud based on a node running on master branch i see this:

![image](https://user-images.githubusercontent.com/2178292/173104201-731a8fd9-4ec2-4ae7-9aba-2a3d04855bda.png)

which is very very close between agent to cloud results

But if i do same process for a node on this branch i see this:

![image](https://user-images.githubusercontent.com/2178292/173104702-4e8f74ff-aa48-4009-8737-b347c5630026.png)

Which is not as good a match and seems to show the agent having a bias towards a lower p-value. 

@MrZammler this is feeling a little similar to what we saw the first time around - any thoughts?