@andrewm4894 2022-06-10
@ktsaou is there any way to version this endpoint somehow such that would be easier to test different algo "generations" on the same agent. Might be a lot of extra work though maybe.

Basically we need to validate the results returned against whats actually live on master and ideally being able to just do this on same node would be best. 

Or if we just had a static range of test input datasets and expected outputs like @ilyam8 suggested that would also maybe be best way forward.

Or for now we can just use the `run_on` param that we can pass to cloud to tell it to run on `cloud` or `agent` and just keep comparing the `agent` results here to `cloud` but its a bit slower and painful this way.