@andrewm4894 2022-06-10
maybe even moving this stuff to something like `api/v1/metric_correlations_tmp` so that we could compare results of `api/v1/metric_correlations` to `api/v1/metric_correlations_tmp` on the same agent. 

But no idea if such an approach would be impossible given the implementation details so i'm not sure on best approach.