@andrewm4894 2022-06-10
anyway its easy enough to just repeat the notebook scatter plot approach each time we have changes to test on this branch