@ktsaou 2022-06-10
@andrewm4894 something is wrong here. The algorithm currently on master, requests the same number of points for both the baseline and the highlight.

Since you are saying that the cloud matches closes the algorithm currently on master, I understand that BOTH of them have the same problem: The baseline is 4x the window but the number of points are the same!

On my version, I fixed that. The baseline has 4 times more points than the hightlight, so that the duration of each point remains the same.

Other than that, the algorithm should be the same.