@andrewm4894 2022-06-10
After discussion with @ktsaou i am in agreement that what we have now is probably more robust to certain edge cases than even our existing cloud implementation (the part that gets the data to the python svc) such that my scatter plot approach above is not as valid as i had thought. 

So after qualitative testing of this i am seeing good results (and maybe even less false positives, but just a feeling) and so happy that we should merge this.