@andrewm4894 2022-06-10
actually - one last test I want to do is validate things a little using [this notebook](https://colab.research.google.com/drive/1PH0loJvWJFCgqD0vzbI2RzQdvzlfu1ck?usp=sharing) which basically calculates `p_api`, `p_py`, `p_staging` and compares them all.