@andrewm4894 2022-06-10
from [this notebook](https://colab.research.google.com/drive/1PH0loJvWJFCgqD0vzbI2RzQdvzlfu1ck?usp=sharing)

`p_py` vs `p_cloud` seems to match very well:

![image](https://user-images.githubusercontent.com/2178292/173124561-1e4f46bd-f9f3-48b1-aa9b-02cd271c6ecb.png)

whereas `p_py` vs `p_agent` still matches ok but not as well: 

![image](https://user-images.githubusercontent.com/2178292/173124701-e1762e73-c4b0-410f-bdbf-7f05ccd6fc5b.png)

But tbh its close enough and probalby just accounts for some precision we have traded off for performance on with this branch. 

I am happy to merge. will approve.