@ktsaou 2022-06-10
Valgrind is reporting that there are some uninitialized data that are used in kstwo. Like not filling the arrays completely. I will try to understand what is happenning tomorrow.

Btw, I think the current master also has a memory leak which I fixed. 

Please don't merge until I say so.