@ktsaou 2022-06-11
The previous code (currently in master) has 2 additional issues:

### Memory leak
This allocation is never freed:
https://github.com/netdata/netdata/blob/6e3144683a73a2024d51425b20ecfd569034c858/database/metric_correlations.c#L154

### not computed dimensions were included in the results

This loop:
https://github.com/netdata/netdata/blob/6e3144683a73a2024d51425b20ecfd569034c858/database/metric_correlations.c#L150

And this loop:
https://github.com/netdata/netdata/blob/6e3144683a73a2024d51425b20ecfd569034c858/database/metric_correlations.c#L186

Both loops do not check if the dimensions were actually selected by the query engine. When dimensions were not selected by the query engine the `kstwo` algorithm was actually run against random memory garbage that happened to be at the position of these dimensions in the results.